// src/common/types/text.types.ts
export enum FONT_WEIGHTS {
  thin = 'Roboto-Thin',
  light = 'Roboto-Light',
  regular = 'Roboto-Regular',
  medium = 'Roboto-Medium',
  bold = 'Roboto-Bold',
}

export enum FONT_SIZES {
  font40 = 40,
  font24 = 24,
  font22 = 22,
  font20 = 20,
  font18 = 18,
  font16 = 16,
  font14 = 14,
  font12 = 12,
}